python_math_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
sympy==1.12
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
pandas==2.0.2
python-multipart==0.0.6
pillow==10.0.0
pytesseract==0.3.10
openai==1.3.0
python-dotenv==1.0.0
redis==5.0.0




